Code refactoring

+++ src/object.cpp
@@ -48,7 +48,7 @@ void Object::Render()
 		// We are passing vertices ...
 		glEnableClientState(GL_VERTEX_ARRAY);
         glEnableClientState(GL_COLOR_ARRAY);
-		if (textureNumber!=-1)
+		if (textureNumber!=(GLuint)-1)
 		{
 			glEnable( GL_TEXTURE_2D );
 			glEnableClientState(GL_TEXTURE_COORD_ARRAY);
